requirements.txt
# Core AI and Language Processing
google-generativeai>=1.0.0
openai-whisper>=20230918
vosk>=0.3.45
SpeechRecognition>=3.10.0

# Text-to-Speech
pyttsx3>=2.90
gTTS>=2.3.0

# Desktop Automation and GUI
pyautogui>=0.9.54
PyQt5>=5.15.10
pystray>=0.19.4
pillow>=10.0.0

# Computer Vision and Screen Capture
opencv-python>=4.8.0
numpy>=1.24.0
screeninfo>=0.8

# Keyboard and Mouse Control
pynput>=1.7.6
keyboard>=0.13.5
mouse>=0.7.1

# Audio Processing
pyaudio>=0.2.11
sounddevice>=0.4.6
librosa>=0.10.0

# Memory and Context Management
chromadb>=0.4.15
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# File and Data Processing
python-docx>=0.8.11
PyPDF2>=3.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.0

# Configuration and Logging
python-dotenv>=1.0.0

# System Integration
psutil>=5.9.5
pywin32>=306;platform_system=='Windows'
AppKit>=0.2.8;platform_system=='Darwin'
python3-xlib>=0.15;platform_system=='Linux'

# MCP (Model Context Protocol) Support
websockets>=12.0
aiohttp>=3.9.0
dataclasses-json>=0.6.0

# Additional dependencies for enhanced features
asyncio-timeout>=4.0.0
typing-extensions>=4.0.0
uuid>=1.30
pathlib>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Enhanced stability
setuptools>=68.0.0